update useToken to action in admin

+++ apiServer/Controllers/UserController.cs
@@ -145,8 +145,23 @@ namespace apiServer.Controllers
             if (userRequest == null)
                 return StatusCode(StatusCodes.Status400BadRequest, "Thiếu dữ liệu để tạo tài khoản!");
 
+            if (userRequest.Password.Length < 6)
+                return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu tối thiểu 6 ký tự");
+
             if (userRequest.Password != userRequest.ConfirmPassword)
                 return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu không khớp!");
+
+            // Kiểm tra định dạng email
+            if (!(new Regex(@"^[^@\s]+@[^@\s]+\.[^@\s]+$").IsMatch(userRequest.Email)))
+            {
+                return StatusCode(StatusCodes.Status400BadRequest, "Email không đúng định dạng.");
+            }
+
+            // Kiểm tra định dạng số điện thoại
+            if (!(new Regex(@"^\d{10}$").IsMatch(userRequest.PhoneNumber)))
+            {
+                return StatusCode(StatusCodes.Status400BadRequest, "PhoneNumber không đúng định dạng.");
+            }
 
             var user = await _context.User.FirstOrDefaultAsync(u => u.Username == userRequest.Username);
             if (user != null)
@@ -192,8 +207,23 @@ namespace apiServer.Controllers
             if (userRequest == null)
                 return StatusCode(StatusCodes.Status400BadRequest, "Thiếu dữ liệu để tạo tài khoản!");
 
+            if (userRequest.Password.Length < 6)
+                return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu tối thiểu 6 ký tự");
+
+
             if (userRequest.Password != userRequest.ConfirmPassword)
                 return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu không khớp!");
+
+            if (!(new Regex(@"^[^@\s]+@[^@\s]+\.[^@\s]+$").IsMatch(userRequest.Email)))
+            {
+                return StatusCode(StatusCodes.Status400BadRequest, "Email không đúng định dạng.");
+            }
+
+            // Kiểm tra định dạng số điện thoại
+            if (!(new Regex(@"^\d{10}$").IsMatch(userRequest.PhoneNumber)))
+            {
+                return StatusCode(StatusCodes.Status400BadRequest, "PhoneNumber không đúng định dạng.");
+            }
 
             var user = await _context.User.FirstOrDefaultAsync(u => u.Username == userRequest.Username);
             if (user != null)
@@ -360,6 +390,11 @@ namespace apiServer.Controllers
         [HttpPost("requestPasswordReset")]
         public async Task<IActionResult> RequestPasswordReset([FromQuery] string email)
         {
+            if (!(new Regex(@"^[^@\s]+@[^@\s]+\.[^@\s]+$").IsMatch(email)))
+            {
+                return StatusCode(StatusCodes.Status400BadRequest, "Email không đúng định dạng.");
+            }
+
             // generate activate account code
             string code = UserService.GenerateVerificationCode();
 
@@ -393,7 +428,7 @@ namespace apiServer.Controllers
             var mailContent = new MailContent();
             mailContent.To = user.Email;
             mailContent.Subject = "Blog Website";
-            mailContent.Body = "<h1>Chào bạn, Đây là mã xác nhận để thay đổi mật khẩu. Lưu ý mã chỉ được sử dụng duy nhất 1 lần!</h1><br/>" + code;
+            mailContent.Body = "<p>Chào bạn, Đây là mã xác nhận để thay đổi mật khẩu. Lưu ý mã chỉ được sử dụng duy nhất 1 lần!</p><br/>" +"<b>" +code+"</b>";
             await MailService.SendMail(mailContent);
             return Ok("Chúng tôi đã gởi mã xác nhận đến email tài khoản của bạn!");
         }
@@ -412,7 +447,7 @@ namespace apiServer.Controllers
 
             if (changePwCode == null)
             {
-                return StatusCode(StatusCodes.Status400BadRequest, "Mã xác nhận không hợp lệ!");
+                return StatusCode(StatusCodes.Status400BadRequest, "Mã xác nhận không chính xác!");
             }
 
             user = await _context.User
@@ -425,6 +460,13 @@ namespace apiServer.Controllers
 
             if (user.Email != email)
                 return StatusCode(StatusCodes.Status400BadRequest, "Email không hợp lệ!");
+
+
+            if (newPassword.Length < 6)
+                return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu tối thiểu 6 ký tự");
+            if (newPassword != confirmPassword)
+                return StatusCode(StatusCodes.Status400BadRequest, "Mật khẩu không khớp");
+
             user.Password = UserService.HashPw(newPassword);
             _context.User.Update(user);
             _context.ChangePwCode.Remove(changePwCode);
